Migrate unregisterSyncRoot
