+++ packages/addon/binding.gyp
@@ -18,8 +18,7 @@
         "native-src/placeholders_interface/Planceholders.cpp",
         "native-src/virtual_drive/get_placeholder_state.cpp",
         "native-src/virtual_drive/get_registered_sync_roots/get_registered_sync_roots.cpp",
-        "native-src/virtual_drive/get_registered_sync_roots/get_registered_sync_roots_wrapper.cpp",
-        "native-src/virtual_drive/unregister_sync_root.cpp"
+        "native-src/virtual_drive/get_registered_sync_roots/get_registered_sync_roots_wrapper.cpp"
       ],
       "include_dirs": [
         "include",
