+++ packages/addon/include/virtual_drive/Wrappers.h
@@ -4,11 +4,6 @@
 #include <node_api.h>
 #include <virtual_drive.h>
 
-inline napi_value UnregisterSyncRootWrapper(napi_env env, napi_callback_info args)
-{
-    return NAPI_SAFE_WRAP(env, args, unregister_sync_root_wrapper);
-}
-
 inline napi_value GetRegisteredSyncRootsWrapper(napi_env env, napi_callback_info args)
 {
     return NAPI_SAFE_WRAP(env, args, get_registered_sync_roots_wrapper);
@@ -17,14 +12,4 @@ inline napi_value GetRegisteredSyncRootsWrapper(napi_env env, napi_callback_info
 inline napi_value GetPlaceholderStateWrapper(napi_env env, napi_callback_info args)
 {
     return NAPI_SAFE_WRAP(env, args, get_placeholder_state_wrapper);
-}
-
-inline napi_value UnwatchPathWrapper(napi_env env, napi_callback_info args)
-{
-    return NAPI_SAFE_WRAP(env, args, unwatchPathWrapper);
-}
-
-inline napi_value WatchPathWrapper(napi_env env, napi_callback_info args)
-{
-    return NAPI_SAFE_WRAP(env, args, watchPathWrapper);
 }
